requirements.txt
python-dotenv==1.0.0
click==8.1.3
rich==13.4.2
flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0    
pyngrok
tinydb
black